update for MultiViewTypes in lib
