--- UltimateRecyclerView/ultimaterecyclerview/src/main/java/com/marshalchen/ultimaterecyclerview/multiViewTypes/EnumListBindAdapter.java
@@ -1,11 +1,0 @@
-package com.marshalchen.ultimaterecyclerview.multiViewTypes;
-
-/**
- * Created by yqritc on 2015/04/20.
- */
-public abstract class EnumListBindAdapter<E extends Enum<E>> extends ListBindAdapter {
-
-    public <T extends DataBinder> T getDataBinder(E e) {
-        return getDataBinder(e.ordinal());
-    }
-}
